Add toast success to users admin page
