+++ src/components/users/index.js
@@ -4,6 +4,8 @@ import BaseDashboard from "../dashboard/base";
 
 import "./index.css"
 import BaseUsers from "./base";
+import CustomToaster from "../sharedComponents/toaster/toastify";
+
 
 const Users = () => {
     return (
@@ -11,6 +13,7 @@ const Users = () => {
             <BaseDashboard
                 ComponentToRender={BaseUsers}
             />
+            <CustomToaster />
         </Fragment>
     );
 };
